Refactored topics so you can use anything, not just String.
The concept of "topic" === String is now forgotten. A topic can be anything, and
it is now a template parameter. WithTopic, TopicQueue and related classes have
been refactored according to that. StringWithTopic, StringTopicQueue and related
classes are introduced to cover the string topic idea. A set StringMap...
classes are also introduced, to encourage the one of the new ways of thinking
about topics. As a result of this, patterns are now replaced with functions for
topic filters. You can still use patterns with the PatternFilter class, that
helps translating your set of patterns to a filter function.

WithTopic now supports null in both topic and element. Affected classes are
changed as well, to provide support. Note that filter caching is not available
for null topics.