Change to unchecked exceptions in module state changes.
A module's state change can throw a random exception, just like what
we protect against in MultiThreadedModule. This should be handled by
the framework, but should not be visible in the methods' interfaces.
Thus, methods working with module state changes (including helpers)
are changed to throw unchecked exceptions and also do deal with
(catch) unchecked exceptions (eg. starting a module).